backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
authlib==1.2.1
itsdangerous==2.1.2
python-dotenv==1.0.0
httpx==0.25.2
starlette==0.27.0
pydantic==2.4.2
pydantic[email]